requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
pydantic>=2.5.0
markdown>=3.5.1
gitpython>=3.1.40
watchdog>=3.0.0
